refactor: move PerkCardSkeleton alongside PerkCard in composite/cards
